refactor: common/ slims to the stack vocabulary; generic infra promoted to xmBase
Follow-through of the common/ review:

- event/ and the SkewSymmetric math helper move to xmBase (rxdu/xmBase#25)
  where every family component can use them; mekf6's private duplicate
  SkewSymmetric is deleted in favor of xmbase/math/matrix_utils.hpp.
- eigen_io (Eigen matrix <-> image, the only stb consumer) becomes a
  private implementation detail of map_processing — its one consumer.
  Its gtest moves along and registers with ctest (was previously the
  math_utils suite).
- src/common/ now holds exactly the navigation-stack vocabulary seed:
  types/trajectory.hpp and control/controller_interface.hpp. Growing
  it into the full inter-area contract (state -> map -> decision ->
  plan -> control) is deliberate future design work, scenario-driven.
- third_party/xmBase re-pins to the event+math commit (v0.4.0 material).

Verified: viz-on 22/22, viz-off 13/13, ASan+UBSan 13/13 (WERROR on).
