--- src/common/event/README.md
@@ -1,1 +1,0 @@
-This module was originally developed in the [quickviz](https://github.com/rxdu/quickviz) project.
