+++ src/estimation/CMakeLists.txt
@@ -4,7 +4,7 @@ find_package(Eigen3 3.3 REQUIRED NO_MODULE)
 add_library(estimation STATIC
     src/mekf6.cpp
     src/mekf9.cpp)
-target_link_libraries(estimation PUBLIC
+target_link_libraries(estimation PUBLIC xmotion::xmBase
     Eigen3::Eigen)
 target_include_directories(estimation PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
