+++ src/mapping/map_processing/src/pgm_map.cpp
@@ -10,7 +10,7 @@
 #include "xmnavigation/map_processing/pgm_map.hpp"
 
 #include "xmbase/telemetry/telemetry.hpp"
-#include "xmnavigation/math_utils/eigen_io.hpp"
+#include "eigen_io.hpp"
 
 #include "pnm/pnm.hpp"
